Add style to context menu
