+++ src/scss/ContextMenu.scss
@@ -1,0 +1,22 @@
+.contextmenu {
+  background-color: #ffffff;
+  font-family: "Courier New", Courier, monospace;
+  border-radius: 5px;
+  padding: 10px;
+  margin: 10px;
+  max-width: 600px;
+  -webkit-box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.85);
+  -moz-box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.85);
+  box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.85);
+}
+
+.contextmenu_item {
+  font-size: 13px;
+  padding: 13px 0;
+  text-align: left;
+  font-weight: 500;
+  color: #909dc2;
+  cursor: pointer;
+  position: relative;
+  border-bottom: 1px solid #e0e2e9;
+}
